# docs/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "RVC2API Backend",
    "description": "Modernized backend API for RV-C CANbus monitoring and control",
    "version": "2.0.0"
  },
  "paths": {
...
    "/api/can/send": {
      "post": {
        "tags": [
          "can"
        ],
        "summary": "Send raw CAN message",
        "description": "Send a raw CAN message to the specified interface.",
        "operationId": "send_raw_message_api_can_send_post",
        "parameters": [
          {
            "name": "arbitration_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Arbitration Id"
            }
          },
          {
            "name": "data",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Data"
            }
          },
          {
            "name": "interface",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Interface"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Send operation result",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Send Raw Message Api Can Send Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/can/recent": {
      "get": {
        "tags": [
          "can"
        ],
        "summary": "Get recent CAN messages",
        "description": "Return recent CAN messages captured on the bus.",
        "operationId": "get_recent_can_messages_api_can_recent_get",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 100,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of recent CAN messages with metadata",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "additionalProperties": true
                  },
                  "title": "Response Get Recent Can Messages Api Can Recent Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}